+++ runtxtserver.py
@@ -8,7 +8,7 @@
 
 import time
 import zmq
-import tempfile
+from zeppy import z_runners 
 
 context = zmq.Context()
 socket = context.socket(zmq.REP)
@@ -18,18 +18,18 @@ import witheppy
 import eppy
 import witheppy.runandget as runandget
 
-def zeppy_runandget(idftxt, wfiletxt, getdict):
-    with tempfile.TemporaryDirectory() as tmpdir:
-        idf_temp_file = f'{tmpdir}/a.idf'
-        open(idf_temp_file, 'w').write(idftxt)
-        wfile_temp_file = f'{tmpdir}/a.epw'
-        open(wfile_temp_file, 'w').write(wfiletxt)
-        print("saved temp files")
-        idf_temp = eppy.openidf(idf_temp_file, epw=wfile_temp_file)
-        fullresult = runandget.anon_runandget(idf_temp, getdict)
-        print(fullresult)
-    return fullresult
-
+# def zeppy_runandget(idftxt, wfiletxt, getdict):
+#     with tempfile.TemporaryDirectory() as tmpdir:
+#         idf_temp_file = f'{tmpdir}/a.idf'
+#         open(idf_temp_file, 'w').write(idftxt)
+#         wfile_temp_file = f'{tmpdir}/a.epw'
+#         open(wfile_temp_file, 'w').write(wfiletxt)
+#         print("saved temp files")
+#         idf_temp = eppy.openidf(idf_temp_file, epw=wfile_temp_file)
+#         fullresult = runandget.anon_runandget(idf_temp, getdict)
+#         print(fullresult)
+#     return fullresult
+#
 while True:
     #  Wait for next request from client
     message = socket.recv_pyobj()
@@ -37,7 +37,7 @@ while True:
     
     #  Do some 'work'
     idftxt, wfiletxt, getdict = message
-    fullresult = zeppy_runandget(idftxt, wfiletxt, getdict)
+    fullresult = z_runners.zeppy_runandget(idftxt, wfiletxt, getdict)
 
     #  Send reply back to client
     socket.send_pyobj(fullresult)
